split convertions
